Refactor monoid skiplist test to generic summarisable storage test
